+++ CHANGES.md
@@ -2,8 +2,9 @@
 Released February ??, 2023
 
 - Applied dialect code and test changes for compatibility with SQLAlchemy 2.0. This
-version of the dialect requires SQLAlchemy 2.0, so to work with earlier versions of
-SQLAlchemy use `pip install sqlalchemy-cockroachdb<2.0.0`
+  version of the dialect requires SQLAlchemy 2.0, so to work with earlier versions of
+  SQLAlchemy use `pip install sqlalchemy-cockroachdb<2.0.0`
+- Stopped sending telemetry data during startup.
 
 # Version 1.4.4
 Released September 9, 2022